requirements.txt
altgraph==0.17
astroid==2.8.0
cffi==1.14.6
colorama==0.4.4
comtypes==1.1.10
future==0.18.2
isort==5.9.3
keyboard==0.13.5
lazy-object-proxy==1.6.0
mccabe==0.6.1
pefile==2021.9.3
platformdirs==2.3.0
pycparser==2.20
pygame==2.0.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pypiwin32==223
pyttsx3==2.90
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
sounddevice==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
vosk==0.3.31
word2number==1.1
wrapt==1.12.1
